# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateReviewDto": {
        "properties": {
          "comment": {
            "description": "Optional - a score alone is a review too. An empty text is stored as no text.",
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "internshipYear": {
            "description": "Buddhist-era year; the service rejects a year after next year.",
            "minimum": 2540,
            "type": "number"
          },
          "position": {
            "description": "Job title during the internship, e.g. \"Frontend Developer Intern\".",
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          },
          "score": {
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "score"
        ],
        "type": "object"
      },
      "DeletedDto": {
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "deleted"
        ],
        "type": "object"
      },
      "PlaceDetailDto": {
        "properties": {
          "averageScore": {
            "description": "Average review score rounded to one decimal; 0 without reviews.",
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "dailyAllowanceSatang": {
            "description": "Daily allowance in satang; 0 = none.",
            "type": "number"
          },
          "distanceKm": {
            "description": "From the university, or from nearLat/nearLng when the list was asked with them.",
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "myReviewId": {
            "description": "The caller's own review of this place, if any.",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "description": "Things to watch out for and benefits.",
            "type": "string"
          },
          "province": {
            "description": "Thai province name without the word \"จังหวัด\".",
            "type": "string"
          },
          "rank": {
            "description": "Position in the ranking of the whole directory (1 = best).",
            "type": "number"
          },
          "reviewCount": {
            "type": "number"
          },
          "reviews": {
            "description": "Newest first.",
            "items": {
              "$ref": "#/components/schemas/ReviewViewDto"
            },
            "type": "array"
          },
          "scoreDistribution": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScoreDistributionDto"
              }
            ],
            "description": "Number of reviews per score."
          },
          "tags": {
            "description": "Keys from GET /api/v1/internship-places/tags (preset keys or users' own words).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "workHours": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "province",
          "latitude",
          "longitude",
          "dailyAllowanceSatang",
          "workHours",
          "notes",
          "tags",
          "averageScore",
          "reviewCount",
          "rank",
          "distanceKm",
          "createdAt",
          "updatedAt",
          "scoreDistribution",
          "reviews",
          "myReviewId"
        ],
        "type": "object"
      },
...
      "ReviewViewDto": {
        "properties": {
          "comment": {
            "description": "null when the reviewer gave a score only.",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "internshipYear": {
            "description": "Buddhist-era year of the internship.",
            "nullable": true,
            "type": "number"
          },
          "isMine": {
            "description": "Whether the caller wrote this review.",
            "type": "boolean"
          },
          "personCode": {
            "description": "Reviewer's student or staff code - returned to staff (place-review:delete:any) only.",
            "nullable": true,
            "type": "string"
          },
          "position": {
            "nullable": true,
            "type": "string"
          },
          "score": {
            "description": "1-5 stars.",
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "score",
          "comment",
          "position",
          "internshipYear",
          "isMine",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ScoreDistributionDto": {
        "properties": {
          "1": {
            "type": "number"
          },
          "2": {
            "type": "number"
          },
          "3": {
            "type": "number"
          },
          "4": {
            "type": "number"
          },
          "5": {
            "type": "number"
          }
        },
        "required": [
          "1",
          "2",
          "3",
          "4",
          "5"
        ],
        "type": "object"
      },
      "UpdatePlaceDto": {
        "properties": {
          "dailyAllowanceSatang": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "notes": {
            "maxLength": 600,
            "minLength": 5,
            "type": "string"
          },
          "province": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "tags": {
            "items": {
              "maxLength": 40,
              "minLength": 2,
              "type": "string"
            },
            "type": "array"
          },
          "workHours": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
      "UpdateReviewDto": {
        "properties": {
          "comment": {
            "description": "Optional - a score alone is a review too. An empty text is stored as no text.",
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "internshipYear": {
            "minimum": 2540,
            "type": "number"
          },
          "position": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          },
          "score": {
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "csmju_internship_directory_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "ระบบสถานที่ฝึกงาน/สหกิจศึกษา และรีวิว — ทุก endpoint ใต้ /api/v1 ต้องมี Core Hub access token (Authorization: Bearer หรือคุกกี้ session ที่ /auth/callback ตั้งให้) และตอบใน envelope { success, data, meta? }",
    "title": "CSMJU Internship Directory API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/internship-places/{id}": {
      "delete": {
        "operationId": "InternshipPlacesController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "internship-places"
        ]
      },
      "get": {
        "operationId": "InternshipPlacesController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlaceDetailDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "internship-places"
        ]
      },
      "patch": {
        "operationId": "InternshipPlacesController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlaceDetailDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "internship-places"
        ]
      }
    },
    "/api/v1/internship-places/{placeId}/reviews": {
      "post": {
        "operationId": "PlaceReviewsController_create",
        "parameters": [
          {
            "in": "path",
            "name": "placeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReviewViewDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "place-reviews"
        ]
      }
    },
    "/api/v1/internship-places/{placeId}/reviews/{reviewId}": {
      "delete": {
        "operationId": "PlaceReviewsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "placeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "place-reviews"
        ]
      },
      "patch": {
        "operationId": "PlaceReviewsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "placeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReviewViewDto"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "place-reviews"
        ]
      }
    },
...
  },
  "servers": [],
...
}